+++ CHANGELOG.md
@@ -3,6 +3,7 @@ This is an overview of major changes. Refer to the git repository for a full log
 Version 0.1.21
 -------------
 - Fixed #208 - Thousands separators for N boxes and N selected boxes (@quicklizard99)
+- Added #205 - A 'Reload template' button (@quicklizard99)
 - Fixed #198 - Opening both .inselect files and images by drag-drop is broken (@quicklizard99)
 - Fixed #48 - Zooming should follow mouse position (@quicklizard99)